LexiDesk/requirements.txt
numpy
pandas
scikit-learn
joblib
tqdm
sentencepiece
torch
torchvision
sklearn_srfsuite